moved all related to service payment to serviceJourney
