+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/service/ServiceJourneyService.java
@@ -16,4 +16,7 @@ public interface ServiceJourneyService {
 
     CcdCallbackResponse serviceDecisionRefusal(CaseDetails caseDetails, String authorisation)
         throws ServiceJourneyServiceException;
+
+    Map<String, Object> setupConfirmServicePaymentEvent(CaseDetails caseDetails)
+        throws ServiceJourneyServiceException;
 }
